fixed prefix in type arg bounds
